Refactor: Fix import and file name typos in RootRoutes.tsx and Routes.tsx
